fix(run): fix package resolution
