+++ src/utils/get-config.ts
@@ -1,17 +1,10 @@
 import fs from 'fs';
 import path from 'path';
 import { Config } from '../types/config.types';
+import { getRootDir } from './get-root-dir';
 
 export function getConfig(): Config {
-  let isRoot = false;
-  let rootDir = process.cwd();
-  while (!isRoot) {
-    isRoot = fs.existsSync(`${rootDir}/.git`);
-    if (!isRoot) {
-      rootDir = `${rootDir}/..`;
-    }
-  }
-
+  const rootDir = getRootDir();
   const packageJSON = JSON.parse(fs.readFileSync(`${rootDir}/package.json`, 'utf8'));
   const config = packageJSON.mookme as Config;
